create and register enum types
