+++ packages/twenty-server/src/engine/metadata-modules/view/enums/view-filter-group-logical-operator.ts
@@ -1,0 +1,5 @@
+export enum ViewFilterGroupLogicalOperator {
+  AND = 'AND',
+  OR = 'OR',
+  NOT = 'NOT',
+}
