+++ packages/twenty-server/src/engine/metadata-modules/view/enums/view-open-recorn-in.ts
@@ -1,0 +1,4 @@
+export enum ViewOpenRecordIn {
+  SIDE_PANEL = 'SIDE_PANEL',
+  RECORD_PAGE = 'RECORD_PAGE',
+}
